requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
email-validator
python-dotenv
cryptography
requests
